tweaks

+++ src/app/documents/page.tsx
@@ -1,21 +1,22 @@
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
+import { mistral } from "@/lib/mistral"
 import { getMaybeUser } from "@/lib/server/auth"
 import { db } from "@/lib/server/db"
 import { documents } from "@/lib/server/db/tests"
+import { setToast } from "@/lib/server/utils/set-toast"
 import { formatDate } from "@/lib/utils"
-import { desc, eq } from "drizzle-orm"
+import { and, desc, eq } from "drizzle-orm"
 import { revalidatePath } from "next/cache"
 import Link from "next/link"
 import { redirect } from "next/navigation"
+import { after } from "next/server"
 import { DeleteDocumentDialog } from "./delete-document-dialog"
 
 export default async function Page() {
   const user = await getMaybeUser()
 
-  if (!user) {
-    redirect("/login")
-  }
+  if (!user) redirect("/login")
 
   const docs = await db.query.documents.findMany({
     where: eq(documents.userId, user.id),
@@ -25,12 +26,16 @@ export default async function Page() {
 
   async function deleteDocument(formData: FormData) {
     "use server"
-
     const user = await getMaybeUser()
     if (!user) redirect("/login")
     const documentId = formData.get("documentId") as string
-    if (!documentId) return
-    await db.delete(documents).where(eq(documents.id, documentId))
+    if (!documentId) return await setToast({ type: "error", message: "Document not found" })
+    const document = await db.query.documents.findFirst({
+      where: and(eq(documents.id, documentId), eq(documents.userId, user.id)),
+    })
+    if (!document) return await setToast({ type: "error", message: "Document not found" })
+    await db.delete(documents).where(and(eq(documents.id, documentId), eq(documents.userId, user.id)))
+    after(mistral.files.delete({ fileId: document.mistralId }))
     revalidatePath("/documents")
   }
 
@@ -52,9 +57,9 @@ export default async function Page() {
         <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
           {docs.map((doc) => (
             <Card key={doc.id}>
-              <CardHeader className="pb-3">
-                <div className="flex items-start justify-between">
-                  <CardTitle>{formatDate(doc.date)}</CardTitle>
+              <CardHeader>
+                <div className="flex items-start justify-between pt-6">
+                  <CardTitle className="font-bold text-lg">{formatDate(doc.date)}</CardTitle>
                   <DeleteDocumentDialog documentId={doc.id} onDelete={deleteDocument} />
                 </div>
                 <CardDescription>
